feat: implement validation if given product exists

+++ src/products/products.controller.ts
@@ -7,6 +7,7 @@ import {
   Post,
   ParseUUIDPipe,
   Put,
+  NotFoundException,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { UpdateProductDto } from './dto/update-product.dto';
@@ -23,11 +24,18 @@ export class ProductsController {
 
   @Get('/:id')
   getById(@Param('id', new ParseUUIDPipe()) id: string) {
-    return this.productsService.getById(id);
+    const product = this.productsService.getById(id);
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
+    return product;
   }
 
   @Delete('/:id')
   deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
     this.productsService.deleteById(id);
     return { success: true };
   }
@@ -42,6 +50,9 @@ export class ProductsController {
     @Param('id', new ParseUUIDPipe()) id: string,
     @Body() productData: UpdateProductDto,
   ) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
     this.productsService.updateById(id, productData);
     return { success: true };
   }
